// src/gen/endpoints/public-auth/public-auth.zod.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import {
  z as zod
} from 'zod';


/**
 * 新規ユーザー登録を行うAPIです。
 * @summary 新規ユーザー登録
 */
export const registerBody = zod.object({
  "first_name": zod.coerce.string().optional().describe('First Name(名)'),
  "last_name": zod.coerce.string().optional().describe('Last Name(姓)'),
  "email_address": zod.coerce.string().optional().describe('E-mailアドレス'),
  "password": zod.coerce.string().optional().describe('パスワード')
}).describe('新規ユーザー登録のリクエスト')

export const loginBody = zod.object({
  "username": zod.coerce.string().optional().describe('ユーザー名(Email Address)'),
  "password": zod.coerce.string().optional().describe('パスワード')
}).describe('ログインのリクエスト')